Добавить основной код по эвентам

+++ main-service/src/main/java/ru/practicum/main/category/CategoryServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.main.event.EventRepository;
 import ru.practicum.main.exception.BadRequestException;
 import ru.practicum.main.exception.ConflictException;
 import ru.practicum.main.exception.NotFoundException;
@@ -16,7 +17,10 @@ import java.util.List;
 @Transactional
 @AllArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
+    private static final String CATEGORY_NOT_FOUND = "Категория с id=%d не найдена.";
+
     private final CategoryRepository repository;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryDto create(NewCategoryDto dto) {
@@ -31,7 +35,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public CategoryDto update(long id, CategoryDto dto) {
         Category category = repository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Категория с id=" + id + " не найдена."));
+                .orElseThrow(() -> new NotFoundException(String.format(CATEGORY_NOT_FOUND, id)));
         category.setName(dto.getName());
         try {
             return CategoryMapper.toDto(repository.save(category));
@@ -43,8 +47,13 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public void delete(long id) {
         if (!repository.existsById(id)) {
-            throw new NotFoundException("Категория с id=" + id + " не найдена.");
+            throw new NotFoundException(String.format(CATEGORY_NOT_FOUND, id));
         }
+
+        if (eventRepository.existsByCategoryId(id)) {
+            throw new ConflictException("В категории есть события.");
+        }
+
         repository.deleteById(id);
     }
 
@@ -52,7 +61,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional(readOnly = true)
     public CategoryDto getById(long id) {
         Category category = repository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Категория с id=" + id + " не найдена."));
+                .orElseThrow(() -> new NotFoundException(String.format(CATEGORY_NOT_FOUND, id)));
         return CategoryMapper.toDto(category);
     }
